pyproject.toml
[project]
name = "moondict"
version = "0.1.0"
description = "Voice dictation for Linux powered by Moonshine AI — fast, offline, lightweight"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "XaviCode1000", email = "gazadev@users.noreply.github.com" },
]
keywords = ["voice", "dictation", "speech-recognition", "moonshine", "offline", "linux"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Typing :: Typed",
]

dependencies = [
    "moonshine-voice>=0.0.51",
    "sounddevice>=0.5.0",
    "numpy>=2.0",
    "pynput>=1.7.7",
    "pystray>=0.19.5",
    "pillow>=10.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[project.scripts]
moondict = "moondict.__main__:main"

[project.urls]
Homepage = "https://github.com/XaviCode1000/moondict"
Issues = "https://github.com/XaviCode1000/moondict/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moondict"]

# ── Ruff (linter + formatter 2026) ──────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # Ruff-specific
    "TCH",   # flake8-type-checking
    "N",     # pep8-naming
    "PTH",   # flake8-use-pathlib
    "RET",   # flake8-return
    "C4",    # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E741",  # ambiguous variable name
    "N818",  # exception naming
    "TC001",  # move application import into type-checking block (false positive for runtime params)
    "TC002",  # move third-party import into type-checking block (false positive for runtime numpy/sd)
]

[tool.ruff.lint.isort]
known-first-party = ["moondict"]

# ── MyPy (type checking) ────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["sounddevice", "moonshine_voice", "simpleaudio", "pynput", "pynput.keyboard", "pystray", "notify2", "notifypy"]
ignore_missing_imports = true

# ── Pytest ──────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short", "--cov=moondict", "--cov-report=term-missing"]
